// smart-shopping/src/app/models/user.ts
export interface user {
   
    firstName: string;
    lastName: string;
    age: number;
    gender: string;
    contactNumber: number;
    userName: string;
    password: string;
    userType: string;
    status: string;
    secretQuestion1: string;
    secretAnswer1: string;
    secretQuestion2: string;
    secretAnswer2: string;
    secretQuestion3: string;
    secretAnswer3: string;
    finalRewardPoints: number;
}